covert profile bar chart to ES6 modules
